refactor: move convergence detector to experiment/
benchmark/convergence.py is the ranked-metric producer for
architecture-comparison-protocol, not submission code — it got feature work
as recently as 2026-06-21 and every --track-experiment run calls it. With
the submission layers gone, the benchmark/ package held only this module,
so it moves to its consumer.

experiment/ over a new analysis/ package: scripts/analysis/ and
tests/quantumnematode_tests/analysis/ already own that name (the latter
path-injects and tests the former), and metrics/ is taken twice
(agent/metrics.py, plasticity/metrics.py). experiment/ is where the sole
caller and the ResultsMetadata model it populates already live, and the
move adds no new import edge — tracker.py already imports report.dtypes,
convergence.py's only non-stdlib dependency.

No symbol renames. Updates the live path reference in
architecture-comparison-protocol; leaves the identical paths under
openspec/changes/archive/ alone, as those record the path as it was.

Verified: 22 convergence tests pass at the new path, 95 experiment tests
green, pyright byte-identical to the 61-error baseline, ruff clean.
