+++ openspec/specs/architecture-comparison-protocol/spec.md
@@ -49,7 +49,7 @@ The episode budget for the C3 primary cells SHALL be set so that every evaluated
 Two distinct convergence operations are used (they are NOT the same test):
 
 - **Budget-setting convergence** (pre-flight): a run is converged for budget-selection purposes when its trailing-window success rate is stable — last-25-mean success within ±5 percentage points of last-100-mean success. This test selects the C3 budget (it set 1000 episodes for the grid-substrate ranking: the recurrent architectures plateau by 1000ep but not 500ep).
-- **Ranked-metric plateau detection** (analysis): the ranked metric is **`post_convergence_success_rate`** — the full-clear (`COMPLETED_ALL_FOOD`) rate averaged over the post-convergence plateau, where the plateau onset is found by `detect_convergence` ([`packages/quantum-nematode/quantumnematode/benchmark/convergence.py`](../../../../packages/quantum-nematode/quantumnematode/benchmark/convergence.py)). Plateau detection SHALL be **level-agnostic**: convergence (the policy has stopped improving) SHALL be decoupled from the absolute success level, so a converged plateau is detected at any band (e.g. a stable 40% plateau and a stable 95% plateau are both detected). Convergence SHALL be established by **absence of trend** — comparing the final trailing block of runs against the immediately-preceding block of equal size and declaring converged when their mean full-clear rates agree within a band calibrated to block sampling noise (NOT by requiring a low-variance near-homogeneous window, which on binary outcomes is reachable only near 100% and so mis-classifies stable intermediate plateaus as non-converged). The plateau onset SHALL be the start of the final region whose smoothed (rolling-mean) success rate has reached the converged level (excluding the warm-up climb, and not anchoring on a transient touch of the band mid-climb), and the metric averages raw full-clear success from that onset to the end of the run, requiring ≥ 30 total runs. A run still trending at its budget SHALL return no plateau (flagged, not mis-scored).
+- **Ranked-metric plateau detection** (analysis): the ranked metric is **`post_convergence_success_rate`** — the full-clear (`COMPLETED_ALL_FOOD`) rate averaged over the post-convergence plateau, where the plateau onset is found by `detect_convergence` ([`packages/quantum-nematode/quantumnematode/experiment/convergence.py`](../../../../packages/quantum-nematode/quantumnematode/experiment/convergence.py)). Plateau detection SHALL be **level-agnostic**: convergence (the policy has stopped improving) SHALL be decoupled from the absolute success level, so a converged plateau is detected at any band (e.g. a stable 40% plateau and a stable 95% plateau are both detected). Convergence SHALL be established by **absence of trend** — comparing the final trailing block of runs against the immediately-preceding block of equal size and declaring converged when their mean full-clear rates agree within a band calibrated to block sampling noise (NOT by requiring a low-variance near-homogeneous window, which on binary outcomes is reachable only near 100% and so mis-classifies stable intermediate plateaus as non-converged). The plateau onset SHALL be the start of the final region whose smoothed (rolling-mean) success rate has reached the converged level (excluding the warm-up climb, and not anchoring on a transient touch of the band mid-climb), and the metric averages raw full-clear success from that onset to the end of the run, requiring ≥ 30 total runs. A run still trending at its budget SHALL return no plateau (flagged, not mis-scored).
 
 The ranked metric SHALL be `post_convergence_success_rate`, NOT a fixed last-N window mean. The fixed-window choice was deliberately rejected because the evaluated arms can have very different warm-up lengths — the from-scratch spiking and quantum arms have long dead-exploration warm-ups on this lethal cell, so a fixed last-N (e.g. last-25) window would mis-measure the slow-igniting arms relative to the fast learners; ranking on the detected post-convergence plateau is the fair comparison. The full-window-mean (plateau mean over a post-warmup tail) SHALL be retained alongside as an **agreement cross-check**: for a homogeneous-warm-up arm set it MUST agree with the detected `post_convergence_success_rate` within noise, bounding the risk that plateau detection introduces bias. (Overall `success_rate`, which includes the warm-up, is also retained in the per-seed export for reference.) For GA cells the analogue is the evolved-champion full-clear rate over a frozen eval. Convergence is distinct from success — an architecture MAY converge to a low plateau (its ceiling, a valid finding) or a high one; the ranked metric SHALL measure that plateau level faithfully whether it is high or low.
